Articulo por id

+++ src/main/java/com/itgt/pos/manager/ArticuloRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.itgt.pos.model.Articulo;
 
-public interface ArticuloRepository extends JpaRepository<Articulo, Integer>{
+public interface ArticuloRepository extends JpaRepository<Articulo, Long>{
 	
 }
